pyproject.toml

[tool.poetry]
name = "helmsman"
version = "0.1.4"
description = "Helmsman is a test framework for helm"
authors = ["MrLYC <imyikong@gmail.com>"]

readme = "README.md"
homepage = "https://github.com/mrlyc/helmsman"
repository = "https://github.com/mrlyc/helmsman"
documentation = "https://helmsman.readthedocs.io"
keywords = ["helmsman"]
classifiers=[
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    
    "License :: OSI Approved :: MIT License",
    
    "Natural Language :: English",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]


[tool.poetry.urls]
"Bug Tracker" = "https://github.com/mrlyc/helmsman/issues"



[tool.poetry.scripts]
helmsman = 'helmsman.cli:main'


[tool.poetry.dependencies]
python = "<3.11,>=3.7"

PyYAML = "^6.0"
py-linq = "^1.3.0"
jmespath = "^0.10.0"


[tool.poetry.dev-dependencies]
pre-commit = "^2.15.0"
invoke = "^1.6.0"
flake8 = "^3.9.2"
flakehell = "^0.9.0"
flake8-builtins = "^1.5.3"
flake8-blind-except = "^0.2.0"
flake8-logging-format = "^0.6.0"
flake8-bugbear = "^21.9.2"
flake8-annotations = "^2.6.2"
flake8-docstrings = "^1.6.0"
flake8-bandit = "^2.1.2"
darglint = "^1.8.0"
isort = "^5.9.3"
black = "^21.9b0"
safety = "^1.10.3"
mypy = "^0.910"
sphinx = "^4.2.0"
recommonmark = "^0.7.1"
bump2version = "^1.0.1"
pytest = "^6.2.5"
xdoctest = "^0.15.10"
coverage = {version = "^6.0.1", extras = ["toml"]}
pytest-cov = "^3.0.0"
watchdog = {version = "^2.1.6", extras = ["watchmedo"]}
ipython = "^7.30.1"
Faker = "^10.0.0"
pytest-mock = "^3.6.1"

[tool.coverage.paths]
source = ["src", "*/site-packages"]

[tool.coverage.run]
branch = true
source = ["helmsman"]

[tool.coverage.report]
fail_under = 100
exclude_lines = [
  "pragma: no cover",
  "def __repr__",
  "if self.debug",
  "if settings.DEBUG:",
  "raise AssertionError",
  "raise NotImplementedError",
  "if 0:",
  "if __name__ == __main__:"
]
show_missing = true

[tool.coverage.html]
directory = "htmlcov"

[tool.flakehell]
format = "grouped"
max_line_length = 99
show_source = true
docstring-convention = "google"
extended_default_ignore = []

[tool.flakehell.plugins]
pyflakes = ["+*"]
pycodestyle = ["+*"]
mccabe = ["+*"]
"flake8-*" = ["+*"]
darglint = ["+*"]

[tool.flakehell.exceptions."tests/"]
flake8-bandit = ["-S101"]

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 99
known_third_party = ["invoke", "nox"]

[tool.black]
line-length = 99
target-version = ["py37"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
